+++ pages/main_page.py
@@ -46,33 +46,3 @@ class MainPage(BasePage):
             expected_conditions.text_to_be_present_in_element_attribute(
                 MainPageLocators.FEED_LINK, 'class', MainPageLocators.ACTIVE_TEXT))
 
-    @allure.step('кликаем на 1-й ингредиент')
-    def click_ingredient_link(self):
-        self.click_element_by_locator(MainPageLocators.INGREDIENT_LINK)
-        #_sleep(5)
-
-    @allure.step('Проверяем, что открывается карточка деталей')
-    def ingredient_details_is_opened(self):
-        return self.wait_for_presence_of_element(MainPageLocators.DETAILS_OPENED_LINK)
-
-    @allure.step('Проверяем, что появился заголовок Детали ингредиента')
-    def details_title_is_visible(self):
-        return self.wait_for_load_element(MainPageLocators.DETAILS_TITLE_LINK)
-
-    @allure.step('кликаем на крестик')
-    def click_details_close_link(self):
-        self.click_element_by_locator(MainPageLocators.DETAILS_CLOSE_LINK)
-        #_sleep(5)
-
-    @allure.step('кликаем на крестик')
-    def click_details_close_link(self):
-        self.click_element_by_locator(MainPageLocators.DETAILS_CLOSE_LINK)
-        #_sleep(5)
-
-    @allure.step('Проверяем что заголовок Детали ингредиента скрыт')
-    def ingredient_details_is_closed(self):
-        return WebDriverWait(self.driver, 10).until(
-            expected_conditions.invisibility_of_element_located(MainPageLocators.DETAILS_TITLE_LINK))
-
-
-
